+++ dsl/view-dsl/src/main/java/org/occurrent/dsl/view/MaterializedView.java
@@ -17,6 +17,8 @@
 
 package org.occurrent.dsl.view;
 
+import org.occurrent.retry.RetryStrategy;
+
 import java.util.function.Function;
 
 /**
@@ -26,9 +28,15 @@ public interface MaterializedView<E> {
     void update(E event);
 
     default <S, ID> void updateFromRepository(ID id, E event, View<S, E> view, ViewStateRepository<S, ID> repository) {
-        S currentState = repository.findById(id).orElse(view.initialState());
-        S updatedState = view.evolve(currentState, event);
-        repository.save(id, updatedState);
+        updateFromRepository(id, event, view, repository, RetryStrategy.none());
+    }
+
+    default <S, ID> void updateFromRepository(ID id, E event, View<S, E> view, ViewStateRepository<S, ID> repository, RetryStrategy retryStrategy) {
+        retryStrategy.execute(() -> {
+            S currentState = repository.findById(id).orElse(view.initialState());
+            S updatedState = view.evolve(currentState, event);
+            repository.save(id, updatedState);
+        });
     }
 
     static <S, E, ID> MaterializedView<E> create(Function<E, ID> idMapper, View<S, E> view, ViewStateRepository<S, ID> repository) {