+++ dsl/view-dsl/src/main/kotlin/org/occurrent/dsl/view/SpringMongoViewExtensions.kt
@@ -17,6 +17,10 @@
 
 package org.occurrent.dsl.view
 
+import org.occurrent.retry.Backoff.exponential
+import org.occurrent.retry.RetryStrategy
+import org.springframework.dao.DuplicateKeyException
+import org.springframework.dao.OptimisticLockingFailureException
 import org.springframework.data.mongodb.core.MongoOperations
 import org.springframework.data.mongodb.core.findById
 import org.springframework.data.repository.CrudRepository
@@ -34,19 +38,67 @@ inline fun <reified S, E, VIEW_ID : Any> View<S, E>.currentState(mongoOperations
 inline fun <reified S, E, VIEW_ID : Any> View<S, E>.materialized(
     mongoOperations: MongoOperations,
     crossinline deriveViewIdFromEvent: (E) -> VIEW_ID
+): MaterializedView<E> = materialized(mongoOperations, SpringMongoViewConfig.config(), deriveViewIdFromEvent)
+
+inline fun <reified S, E, VIEW_ID : Any> View<S, E>.materialized(
+    mongoOperations: MongoOperations,
+    config: SpringMongoViewConfig = SpringMongoViewConfig.config(),
+    crossinline deriveViewIdFromEvent: (E) -> VIEW_ID
 ): MaterializedView<E> {
     val noopStateConvert = object : StateConverter<S, S & Any> {
         override fun toDTO(viewState: S): S & Any = viewState as (S & Any)
         override fun fromDTO(dto: S & Any): S = dto
     }
-    return materialized(mongoOperations, noopStateConvert, deriveViewIdFromEvent)
+    return materialized(mongoOperations, noopStateConvert, config, deriveViewIdFromEvent)
 }
 
 inline fun <S_VIEW, reified S_DTO : Any, E, VIEW_ID : Any> View<S_VIEW, E>.materialized(
     mongoOperations: MongoOperations,
     converter: StateConverter<S_VIEW, S_DTO>,
     crossinline deriveViewIdFromEvent: (E) -> VIEW_ID
+): MaterializedView<E> = materialized(mongoOperations, converter, SpringMongoViewConfig.config(), deriveViewIdFromEvent)
+
+inline fun <S_VIEW, reified S_DTO : Any, E, VIEW_ID : Any> View<S_VIEW, E>.materialized(
+    mongoOperations: MongoOperations,
+    converter: StateConverter<S_VIEW, S_DTO>,
+    config: SpringMongoViewConfig = SpringMongoViewConfig.config(),
+    crossinline deriveViewIdFromEvent: (E) -> VIEW_ID
 ): MaterializedView<E> {
+    val (duplicateKeyHandling, optimisticLockingHandling) = config
+    val retryStrategy: RetryStrategy = RetryStrategy.retry()
+        .let { rs ->
+            if (optimisticLockingHandling is OptimisticLockingHandling.Retry) {
+                rs.backoff(exponential(optimisticLockingHandling.initial, optimisticLockingHandling.max, optimisticLockingHandling.multiplier))
+            } else {
+                rs
+            }
+        }
+        .retryIf { e ->
+            when (e) {
+                is OptimisticLockingFailureException -> optimisticLockingHandling is OptimisticLockingHandling.Retry
+                else -> false
+            }
+        }
+        .onError { e ->
+            when (e) {
+                is DuplicateKeyException -> if (duplicateKeyHandling is DuplicateKeyHandling.Ignore) {
+                    duplicateKeyHandling.onDuplicateKeyException(e)
+                }
+
+                is OptimisticLockingFailureException -> when (optimisticLockingHandling) {
+                    is OptimisticLockingHandling.Ignore -> {
+                        optimisticLockingHandling.onOptimisticLockingFailureException(e)
+                    }
+
+                    is OptimisticLockingHandling.Retry -> {
+                        optimisticLockingHandling.onOptimisticLockingFailureException(e)
+                    }
+
+                    else -> {}
+                }
+            }
+        }
+
     val viewStateRepository = object : ViewStateRepository<S_VIEW, VIEW_ID> {
         override fun findById(id: VIEW_ID): Optional<S_VIEW & Any> = Optional.ofNullable(mongoOperations.findById(id, S_DTO::class.java))
             .map { dto -> converter.fromDTO(dto) }
@@ -60,7 +112,17 @@ inline fun <S_VIEW, reified S_DTO : Any, E, VIEW_ID : Any> View<S_VIEW, E>.mater
     val view = this
     return object : MaterializedView<E> {
         override fun update(event: E) {
-            updateFromRepository(deriveViewIdFromEvent(event), event, view, viewStateRepository)
+            try {
+                updateFromRepository(deriveViewIdFromEvent(event), event, view, viewStateRepository, retryStrategy)
+            } catch (e: DuplicateKeyException) {
+                if (duplicateKeyHandling is DuplicateKeyHandling.Rethrow) {
+                    throw e
+                }
+            } catch (e: OptimisticLockingFailureException) {
+                if (optimisticLockingHandling is OptimisticLockingHandling.Rethrow) {
+                    throw e
+                }
+            }
         }
     }
 }